Implement data-access guard + Refactoring

+++ src/budget/budget.controller.ts
@@ -1,6 +1,19 @@
-import { Body, Controller, Delete, Get, HttpException, HttpStatus, Param, Patch, Post, Query, UseGuards } from '@nestjs/common';
-import { Roles } from 'src/auth/decorator';
-import { JwtGuard } from 'src/auth/guard';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpException,
+  HttpStatus,
+  Param,
+  Patch,
+  Post,
+  Query,
+  Req,
+  UseGuards,
+} from '@nestjs/common';
+import { DataAccessGuard, JwtGuard } from 'src/auth/guard';
+import { CustomRequest } from 'src/user/models/request.models';
 import { BudgetService } from './budget.service';
 import { CreateBudgetDto, UpdateBudgetDto } from './dto';
 
@@ -10,24 +23,38 @@ export class BudgetController {
   constructor(private readonly budgetService: BudgetService) {}
 
   @Get('')
-  async getBudgets(@Query() query: Object) {
+  @UseGuards(DataAccessGuard)
+  async getBudgets(@Query() query: any, @Req() req: CustomRequest) {
     try {
-      return await this.budgetService.getBudgetsWithParams(query);
+      const user = req.userObj;
+      if (user.role === 'ADMIN') {
+        return await this.budgetService.getBudgetsWithParams(query);
+      } else {
+        query.userId = user.id;
+        return await this.budgetService.getBudgetsWithParams(query);
+      }
     } catch (error) {
+      console.log(error);
       throw new HttpException('No Budgets found', HttpStatus.NOT_FOUND);
     }
   }
 
   @Get('/:id')
-  async getBudget(@Param('id') id: string) {
+  @UseGuards(DataAccessGuard)
+  async getBudget(@Param('id') id: string, @Req() req: CustomRequest) {
     try {
+      const user = req.userObj;
       const budget = await this.budgetService.getBudgetById(id);
       if (!budget) {
         throw new Error('Budget not found');
       }
+      if (user.id !== budget.userId && user.role !== 'ADMIN') {
+        throw new Error('Budget belong to another user');
+      }
       return budget;
     } catch (error) {
-      throw new HttpException('Budget not found', HttpStatus.NOT_FOUND);
+      console.log(error);
+      throw new HttpException(error.message, HttpStatus.NOT_FOUND);
     }
   }
 
@@ -36,38 +63,52 @@ export class BudgetController {
     try {
       return await this.budgetService.addBudget(createBudgetDto);
     } catch (error) {
-      throw new HttpException(
-        'Unexpected error occured',
-        HttpStatus.BAD_REQUEST,
-      );
+      console.log(error);
+      throw new HttpException(error, HttpStatus.BAD_REQUEST);
     }
   }
 
   @Patch('/:id')
+  @UseGuards(DataAccessGuard)
   async updateBudget(
     @Param('id') id: string,
     @Body() updateBudgetDto: UpdateBudgetDto,
+    @Req() req: CustomRequest,
   ) {
     try {
+      const user = req.userObj;
+      const budget = await this.budgetService.getBudgetById(id);
+
+      if (!budget) {
+        throw new Error('Budget not found');
+      }
+
+      if (user.id !== budget.userId && user.role !== 'ADMIN') {
+        throw new Error('Budget belong to another user');
+      }
       return await this.budgetService.updateBudget(id, updateBudgetDto);
     } catch (error) {
-      throw new HttpException(
-        'Unexpected error occured',
-        HttpStatus.BAD_REQUEST,
-      );
+      console.log(error);
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
   }
 
   @Delete('/:id')
-  async deleteBudget(@Param('id') id: string) {
+  @UseGuards(DataAccessGuard)
+  async deleteBudget(@Param('id') id: string, @Req() req: CustomRequest) {
     try {
-      const budget = await this.budgetService.removeBudget(id);
+      const user = req.userObj;
+      const budget = await this.budgetService.getBudgetById(id);
+
       if (!budget) {
         throw new Error('Budget not found');
       }
-      return budget;
+      if (user.id !== budget.userId && user.role !== 'ADMIN') {
+        throw new Error('Budget belong to another user');
+      }
+      return await this.budgetService.removeBudget(id);
     } catch (error) {
-      throw new HttpException('Budget not found', HttpStatus.NOT_FOUND);
+      throw new HttpException(error.message, HttpStatus.NOT_FOUND);
     }
   }
 }
